Let compatLibVR depend on compatLibVQ
