+++ compatLib/compatLibVQ/build.gradle
@@ -10,5 +10,4 @@ addFrameworkJar('framework-10.jar')
 
 dependencies {
     implementation projects.compatLib
-    implementation projects.compatLib.compatLibVR
 }
