+++ compatLib/compatLibVQ/src/main/java/app/lawnchair/compatlib/ten/QuickstepCompatFactoryVQ.java
@@ -1,11 +1,16 @@
 package app.lawnchair.compatlib.ten;
 
+import android.app.IApplicationThread;
+import android.window.IRemoteTransition;
+import android.window.RemoteTransition;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import app.lawnchair.compatlib.ActivityManagerCompat;
 import app.lawnchair.compatlib.ActivityOptionsCompat;
-import app.lawnchair.compatlib.eleven.QuickstepCompatFactoryVR;
+import app.lawnchair.compatlib.QuickstepCompatFactory;
+import app.lawnchair.compatlib.RemoteTransitionCompat;
 
-public class QuickstepCompatFactoryVQ extends QuickstepCompatFactoryVR {
+public class QuickstepCompatFactoryVQ extends QuickstepCompatFactory {
 
     @NonNull
     @Override
@@ -17,5 +22,19 @@ public class QuickstepCompatFactoryVQ extends QuickstepCompatFactoryVR {
     @Override
     public ActivityOptionsCompat getActivityOptionsCompat() {
         return new ActivityOptionsCompatVQ();
+    }
+
+    @NonNull
+    @Override
+    public RemoteTransitionCompat getRemoteTransitionCompat() {
+        return new RemoteTransitionCompat() {
+            @Override
+            public RemoteTransition getRemoteTransition(
+                    @NonNull IRemoteTransition remoteTransition,
+                    @Nullable IApplicationThread appThread,
+                    @Nullable String debugName) {
+                return new RemoteTransition(remoteTransition, appThread, debugName);
+            }
+        };
     }
 }
