+++ compatLib/src/main/java/app/lawnchair/compatlib/RemoteTransitionCompat.java
@@ -1,10 +1,10 @@
 package app.lawnchair.compatlib;
 
-import android.annotation.NonNull;
-import android.annotation.Nullable;
 import android.app.IApplicationThread;
 import android.window.IRemoteTransition;
 import android.window.RemoteTransition;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public abstract class RemoteTransitionCompat {
